requirements.txt
# SynergyFrame 依赖库
numpy>=1.19.0
opencv-python>=4.5.0
torch>=1.10.0
torchvision>=0.11.0
pillow>=9.0.0
matplotlib>=3.4.0
diffusers>=0.24.0
transformers>=4.30.0
rembg>=2.0.0
safetensors>=0.3.0
accelerate>=0.20.0
xformers>=0.0.20
timm>=0.6.0
huggingface-hub>=0.14.0
tqdm>=4.65.0
scikit-image>=0.19.0
